:recycle: move code to common module
